agentapp/requirements.txt
uvicorn==0.35.0
aiohttp==3.12.15
quart==0.20.0
quart-auth==0.11.0
quart-rate-limiter==0.12.0
quart-wtforms==1.0.3
redis==6.4.0
email_validator==2.2.0
asyncio==4.0.0
nest-asyncio==1.6.0
quart-wtforms[email]
setuptools==80.9.0
passlib==1.7.4
bcrypt==4.3.0
tzlocal==5.3.1
pyjwt==2.10.1
openai==1.99.1
Werkzeug==3.1.3
asyncio==4.0.0
nest-asyncio==1.6.0
quart-wtforms[email]
setuptools==80.9.0
passlib==1.7.4
bcrypt==4.3.0
tzlocal==5.3.1
pyjwt==2.10.1
langgraph
langchain-mcp-adapters
